@@ -1,27 +1,18 @@
 {
   "author": "Sebastian Pekarek <mail@sebbo.net>",
-  "bin": {
-    "test-cli": "./dist/bin/cli.js"
-  },
   "bugs": {
     "url": "https://github.com/sebbo2002/js-template/issues"
   },
-  "dependencies": {
-    "express": "^4.17.1"
-  },
-  "description": "A template that I use to start new projects and services.",
+  "description": "GitHub Action that creates an annotation with the resulting release type when using semantic-release",
   "devDependencies": {
-    "@amanda-mitchell/semantic-release-npm-multiple": "^2.15.0",
-    "@qiwi/semantic-release-gh-pages-plugin": "^5.2.0",
-    "@sebbo2002/semantic-release-docker": "^1.0.0-develop.1",
     "@semantic-release/changelog": "^6.0.1",
     "@semantic-release/exec": "^6.0.2",
     "@semantic-release/git": "^10.0.1",
-    "@types/express": "^4.17.13",
     "@types/mocha": "^9.0.0",
     "@types/node": "^16.11.10",
     "@typescript-eslint/eslint-plugin": "^5.4.0",
     "@typescript-eslint/parser": "^5.4.0",
+    "@vercel/ncc": "^0.33.4",
     "eslint": "^8.3.0",
     "eslint-plugin-jsonc": "^2.0.0",
     "license-checker": "^25.0.1",
@@ -30,12 +21,10 @@
     "nyc": "^15.1.0",
     "semantic-release-license": "^1.0.3",
     "source-map-support": "^0.5.21",
-    "ts-node": "^10.4.0",
-    "typedoc": "^0.22.10",
-    "typescript": "^4.5.2"
+    "ts-node": "^10.7.0"
   },
   "engines": {
-    "node": ">=12.0.0"
+    "node": ">=14.0.0"
   },
   "files": [
     "/dist"
@@ -49,14 +38,20 @@
     "url": "git+https://github.com/sebbo2002/js-template.git"
   },
   "scripts": {
-    "build": "tsc && chmod +x ./dist/bin/*.js",
+    "build": "ncc build ./src/index.ts -m --license licenses.txt",
     "build-all": "./.github/workflows/build.sh",
     "coverage": "nyc mocha",
-    "develop": "ts-node ./src/bin/start.ts",
-    "license-check": "license-checker --production --summary",
+    "license-check": "license-checker --production --onlyAllow \"MIT;ISC;Apache-2.0\"",
     "lint": "eslint . --ext .ts,.json",
-    "start": "node ./dist/bin/start.js",
     "test": "mocha"
   },
-  "version": "1.0.0"
+  "version": "1.0.0",
+  "dependencies": {
+    "@actions/core": "^1.7.0",
+    "@actions/github": "^5.0.1",
+    "@octokit/rest": "^18.12.0",
+    "@semantic-release/commit-analyzer": "^9.0.2",
+    "@types/semantic-release": "^17.2.3",
+    "cosmiconfig": "^7.0.1"
+  }
 }
